refactor(frontend): Move foldInternalCalls to LintService
